// src/solidity/type.ts
export interface StructLog {
    depth: number,
    error?: string,
    invalidDepthChange: boolean,
    // remaining gas
    gas: number,
    // step cost
    gasCost: number,
    memory: string[],
    memexpand?: string;
    op: string,
    pc: number,
    stack: string[],
    storage: {
        [key: string]: string;
    };
}